Update adsorption.py, bidentate and site selection
Now it is possible to create bidentate adsorbates in which two atoms of the adsorbate not directly connected (e.g., O and O in HCOO) are bonded to the surface sites. 
A general adsorption methodology for bidentate adsorption (auto_construct=False) is introduced, exploiting ase.build.tools.rotation_matrix. 
Now it is possible to select a list of active sites to adsorb the molecules (monodentate and bidentate). This can limit the number of the adsorption configurations, once known the preferred sites of each adsorbate (parameter : sites_names).